document POST activities endpoint (#165)
* documented activities

* improved and documented error handling

* fixed typo

* cleanup

* added errors on create document and create note

* fixed testgit add .git add .

* added unit tests for error handling

* model tests added

* added integration tests for error scenarios. Some cleanup

+++ models/Publications_Tags.js
@@ -1,6 +1,7 @@
 const { Model } = require('objection')
 const parseurl = require('url').parse
 const { Publication } = require('./Publication')
+const { Tag } = require('./Tag')
 
 class Publications_Tags extends Model {
   static get tableName () {
@@ -15,8 +16,16 @@ class Publications_Tags extends Model {
     publicationUrl /*: string */,
     tagId /*: number */
   ) /*: any */ {
+    // check publication
+    if (!publicationUrl) return new Error('no publication')
     let publicationShortId = parseurl(publicationUrl).path.substr(13)
     const publication = await Publication.byShortId(publicationShortId)
+    if (!publication) return new Error('no publication')
+
+    // check tag
+    if (!tagId) return new Error('no tag')
+    const tag = await Tag.byId(tagId)
+    if (!tag) return new Error('no tag')
     // check if already exists
     const result = await Publications_Tags.query().where({
       publicationId: publication.id,
@@ -35,8 +44,17 @@ class Publications_Tags extends Model {
     publicationUrl /*: string */,
     tagId /*: string */
   ) /*: number */ {
+    // check publication
+    if (!publicationUrl) return new Error('no publication')
     let publicationShortId = parseurl(publicationUrl).path.substr(13)
     const publication = await Publication.byShortId(publicationShortId)
+    if (!publication) return new Error('no publication')
+
+    // check tag
+    if (!tagId) return new Error('no tag')
+    const tag = await Tag.byId(tagId)
+    if (!tag) return new Error('no tag')
+
     return await Publications_Tags.query()
       .delete()
       .where({
